// server/src/schemas/user.schema.ts
import { z } from 'zod';

export const createUserSchema = z.object({
    body: z.object({
        name: z.string().min(3, "Name must be at least 3 characters"),
        email: z.string().email({ message: "Invalid email format" }),
        password: z.string().min(8, "Password must be at least 8 characters"),
    }),
});

export type CreateUserInput = z.infer<typeof createUserSchema>['body'];